Update argsort routine to also return sorted values
This doesn't make CPU sorting significantly slower. It does double the
amount of memory though and make WebGPU argsort about 65% slower for 1024
elements (fits in workgroup), however that is expected. It'll be needed
anyway when we support GPU sorting on >1024 elements, and argsort isn't
important enough to have separate code paths.

The end goal btw is to have a primitive like `jax.lax.sort()`, which
sorts a set of arrays lexicographically. This is one step closer to that
since argsort is like sorting (A, I) for index array I.